// src/types/twitchAPI.ts
interface Pagination {
  cursor?: string; // Optional, as it may not exist if there are no more pages
}
// ...
enum Label {
  DrugsIntoxication = "DrugsIntoxication",
  SexualThemes = "SexualThemes",
  ViolentGraphic = "ViolentGraphic",
  Gambling = "Gambling",
  ProfanityVulgarity = "ProfanityVulgarity",
}

// Interface for Content Classification Labels (CCLs)
interface ContentClassificationLabel {
  id: Label;
  is_enabled: boolean;
}

export interface ModifyChannelInformationRequest {
  game_id?: string; // Optional, with specific behavior described for "0" or ""
  broadcaster_language?: string; // Optional, with constraints on the values
  title?: string; // Cannot be an empty string
  delay?: number; // Optional, with additional constraints for Partner status
  tags?: string[]; // Optional, with constraints on the number of tags, their length, and character content
  content_classification_labels?: ContentClassificationLabel[]; // Optional, list of labels for content classification
  is_branded_content?: boolean; // Optional, indicates if the channel has branded content
}
// ...
export interface getChannelSubscriptionsReponse {
  data: Array<{
    broadcaster_id: string; // An ID that identifies the broadcaster.
    broadcaster_login: string; // The broadcaster’s login name.
    broadcaster_name: string; // The broadcaster’s display name.
    gifter_id: string; // The ID of the user that gifted the subscription to the user. Is an empty string if is_gift is false.
    gifter_login: string; // The gifter’s login name. Is an empty string if is_gift is false.
    gifter_name: string; // The gifter’s display name. Is an empty string if is_gift is false.
    is_gift: boolean; // A Boolean value that determines whether the subscription is a gift subscription. Is true if the subscription was gifted.
    plan_name: string; // The name of the subscription.
    tier: "1000" | "2000" | "3000"; // The type of subscription. Possible values are 1000 — Tier 1, 2000 — Tier 2, 3000 — Tier 3.
    user_id: string; // An ID that identifies the subscribing user.
    user_name: string; // The user’s display name.
    user_login: string; // The user’s login name.
  }>;

  // Contains the information used to page through the list of results. The object is empty if there are no more pages left to page through.
  pagination: {
    cursor: string; // The cursor used to get the next or previous page of results. Use the cursor to set the request’s after or before query parameter depending on whether you’re paging forwards or backwards.
  };

  points: number; // The current number of subscriber points earned by this broadcaster. Points are based on the subscription tier of each user that subscribes to this broadcaster. For example, a Tier 1 subscription is worth 1 point, Tier 2 is worth 2 points, and Tier 3 is worth 6 points. The number of points determines the number of emote slots that are unlocked for the broadcaster.

  total: number; // The total number of users that subscribe to this broadcaster.
}

interface Follower {
  followed_at: string; // The UTC timestamp when the user started following the broadcaster
  user_id: string; // An ID that uniquely identifies the user that’s following the broadcaster
  user_login: string; // The user’s login name
  user_name: string; // The user’s display name
}
export interface getChannelFollowersReponse {
  data: Follower[]; // The list of users that follow the specified broadcaster
  pagination: Pagination; // Contains the information used to page through the list of results
  total: number; // The total number of users that follow this broadcaster
}